feat: useAppSettings
